backend/requirements.txt
PyJWT
pyotp
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
pydantic[email]
python-dotenv==1.0.0
qrcode[pil]==7.4.2
python-multipart==0.0.6
slowapi==0.1.9
bcrypt==4.1.1
alembic==1.13.0